+++ toolkit/components/credentialmanagement/IdentityCredentialPromptService.sys.mjs
@@ -157,7 +157,8 @@ export class IdentityCredentialPromptService {
         const data = {
           id: providerIndex,
           icon: iconResult.value,
-          name: displayDomain,
+          name: providerNames[providerIndex],
+          domain: displayDomain,
         };
         providers.push(data);
       }
@@ -544,6 +545,12 @@ export class IdentityCredentialPromptService {
       {}
     );
 
+    let headerIconResult = await this.loadIconFromManifest(
+      providerManifest,
+      BEST_HEADER_ICON_SIZE,
+      "chrome://global/skin/icons/defaultFavicon.svg"
+    );
+
     if (AppConstants.platform === "android") {
       const accounts = [];
 
@@ -565,8 +572,14 @@ export class IdentityCredentialPromptService {
         console.log(data);
       }
 
+      const provider = {
+        name: providerName || displayDomain,
+        domain: displayDomain,
+        icon: headerIconResult,
+      };
+
       const result = {
-        provider: displayDomain,
+        provider,
         accounts,
       };
 
@@ -648,12 +661,6 @@ export class IdentityCredentialPromptService {
       // Add the item to the DOM!
       listBox.append(newItem);
     }
-
-    let headerIconResult = await this.loadIconFromManifest(
-      providerManifest,
-      BEST_HEADER_ICON_SIZE,
-      "chrome://global/skin/icons/defaultFavicon.svg"
-    );
 
     // Create a new promise to wrap the callbacks of the popup buttons
     return new Promise(function (resolve, reject) {